Add creature abstract class.
