+++ src/main/java/Animal.java
@@ -2,22 +2,13 @@ import org.sql2o.*;
 import java.util.ArrayList;
 import java.util.List;
 
-public class Animal {
-  public String name;
-  public int id;
+public class Animal extends Creature {
 
   public Animal(String name) {
     this.name = name;
     this.id = id;
   }
 
-  public String getName() {
-    return name;
-  }
-
-  public int getId() {
-    return id;
-  }
 
   public List<Sighting> getSightings() {
     try(Connection con = DB.sql2o.open()) {
